+++ metricbeat/docs/reference/configuration.asciidoc
@@ -18,6 +18,7 @@ configuration settings, you need to restart {beatname_uc} to pick up the changes
 * <<console-output>>
 * <<configuration-output-ssl>>
 * <<configuration-path>>
+* <<configuration-dashboards>>
 * <<configuration-logging>>
 * <<configuration-processors>>
