+++ packetbeat/docs/reference/configuration.asciidoc
@@ -21,6 +21,7 @@ configuration settings, you need to restart {beatname_uc} to pick up the changes
 * <<console-output>>
 * <<configuration-output-ssl>>
 * <<configuration-path>>
+* <<configuration-dashboards>>
 * <<configuration-logging>>
 * <<configuration-run-options>>
 * <<configuration-processors>>
