+++ winlogbeat/docs/reference/configuration/winlogbeat-options.asciidoc
@@ -358,6 +358,9 @@ include::../../../../libbeat/docs/outputconfig.asciidoc[]
 pass::[<?edit_url https://github.com/elastic/beats/edit/master/libbeat/docs/shared-path-config.asciidoc ?>]
 include::../../../../libbeat/docs/shared-path-config.asciidoc[]
 
+pass::[<?edit_url https://github.com/elastic/beats/edit/master/libbeat/docs/dashboardsconfig.asciidoc ?>]
+include::../../../../libbeat/docs/dashboardsconfig.asciidoc[]
+
 pass::[<?edit_url https://github.com/elastic/beats/edit/master/libbeat/docs/loggingconfig.asciidoc ?>]
 include::../../../../libbeat/docs/loggingconfig.asciidoc[]
